removed a redundant (and wrong) cast.
2008-04-08  Sven Neumann  <sven@gimp.org>

	* plug-ins/common/mng.c (myalloc): removed a redundant (and wrong)
	cast.

svn path=/trunk/; revision=25412

+++ plug-ins/common/mng.c
@@ -220,7 +220,7 @@ myalloc (mng_size_t size)
 {
   gpointer ptr;
 
-  ptr = g_try_malloc ((gulong) size);
+  ptr = g_try_malloc (size);
 
   if (ptr != NULL)
     memset (ptr, 0, size);
